@@ -1,12 +1,12 @@
 import React from 'react';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import { Products } from '../pages/Products';
 import { Cart } from '../pages/Cart';
 import { Home } from '../pages/Home';
 import { Header } from '../components/Header';
 import { Login } from '../pages/Login';
 import { Logout } from '../pages/Logout';
-import RequiredAuthRoute from '../components/RequiredAuthRoute';
+import RequiredAuth from '../components/RequiredAuth';
 import AuthProvider from '../components/AuthProvider';
 
 export const AppRouter = () => {
@@ -14,23 +14,42 @@ export const AppRouter = () => {
     <AuthProvider>
       <Router>
         <Header />
-        <Switch>
-          <RequiredAuthRoute exact path="/">
-            <Home />
-          </RequiredAuthRoute>
-          <RequiredAuthRoute path="/products">
-            <Products />
-          </RequiredAuthRoute>
-          <RequiredAuthRoute path="/cart">
-            <Cart />
-          </RequiredAuthRoute>
-          <Route path="/login">
-            <Login />
-          </Route>
-          <RequiredAuthRoute path="/logout">
-            <Logout />
-          </RequiredAuthRoute>
-        </Switch>
+        <Routes>
+          <Route
+            exact
+            path="/"
+            element={
+              <RequiredAuth>
+                <Home />
+              </RequiredAuth>
+            }
+          />
+          <Route
+            path="/products"
+            element={
+              <RequiredAuth>
+                <Products />
+              </RequiredAuth>
+            }
+          />
+          <Route
+            path="/cart"
+            element={
+              <RequiredAuth>
+                <Cart />
+              </RequiredAuth>
+            }
+          />
+          <Route path="/login" element={<Login />} />
+          <Route
+            path="/logout"
+            element={
+              <RequiredAuth>
+                <Logout />
+              </RequiredAuth>
+            }
+          />
+        </Routes>
       </Router>
     </AuthProvider>
   );
